Force cgo for golang DNS resolver
The default behavior of the golang net package is to use the pure go
implementation which sends DNS requests directly to the servers listed
in resolv.conf.  This breaks when running the docker client inside a
docker container that's 'linked' with other containers given that those
links utilize teh /etc/hosts to function.  This change forces golang's
netdns to use cgo which uses the system c libraries and respects
nsswitch.

Before this change any short name conflict will result in the DNS query
resolving to a remote host instead of the linked container.  For
example, docker will resolve to docker.rc.fas.harvard.edu instead of
the linked container due to the search domain in /etc/resolv.conf.