Restrict channel aliases to 19.03
This will remove the "test" and "stable" aliases for whatever major release comes after 19.03 (so once the next release comes out, "stable" will become a stale alias).

I'm not 100% sure this is the right answer, but the "stable" alias has definitely been a huge source of confusion very recently (and it doesn't have much meaning anymore anyways given the death of the "edge" channel).