Switch over to xtables-legacy when nf_tables module isn't available
PR #461 updated Alpine to 3.19 and made a change to load the nf_tables
kernel module if needed. However, as demonstrated by #463 and #464 this
might break when the host system doesn't have the nf_tables module
available. In that case, we should still try to load the ip_tables
module and symlink /sbin/iptables to xtables-legacy-multi.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>