Ignore the exit code of `modprobe` always
The nature of `modprobe` in this image is that it works via `ip` hacks, but the exit code will always be non-zero because we don't have `/lib/modules` from the host.

The effect of this was that everyone was using `iptables-legacy` (whether it was warranted for them to be doing so or not).