Use UID number rather than username in Dockerfile
Systems configured to disallow running images as root aren't able to run images that use user name string values for the USER because they can't validate that a named user isn't root. To allow this image to run on such systems, use the uid of the user as the value for USER instead of the username.

See: https://github.com/kubernetes/kubernetes/pull/56503

Signed-off-by: Craig Andrews <candrews@integralblue.com>